refactor: move bubble backfill to a lib so can be used by other projects eg LightDAS and be included in the grpc-ingest to execute backfills on restart then pick up tasks from the stream to insure no out of order indexing
