+++ Builder.Dockerfile
@@ -1,11 +1,12 @@
 FROM rust:1.75-bullseye AS builder
 RUN apt-get update -y && \
-    apt-get install -y build-essential make git
-    
+  apt-get install -y build-essential make git
+
 RUN mkdir /rust
 RUN mkdir /rust/bins
 COPY Cargo.toml /rust
 COPY core /rust/core
+COPY backfill /rust/backfill
 COPY das_api /rust/das_api
 COPY digital_asset_types /rust/digital_asset_types
 COPY integration_tests /rust/integration_tests
@@ -19,7 +20,7 @@ COPY blockbuster rust/blockbuster
 WORKDIR /rust
 RUN --mount=type=cache,target=/rust/target,id=das-rust \
   cargo build --release --bins && cp `find /rust/target/release -maxdepth 1 -type f | sed 's/^\.\///' | grep -v "\." ` /rust/bins
-    
+
 FROM rust:1.75-slim-bullseye as final
 COPY --from=builder /rust/bins /das/
 CMD echo "Built the DAS API bins!"
