+++ Proxy.Dockerfile
@@ -4,6 +4,7 @@ RUN cargo install wasm-pack
 RUN mkdir /rust
 COPY ./Cargo.toml /rust
 COPY ./core /rust/core
+COPY ./backfill /rust/backfill
 COPY ./das_api /rust/das_api
 COPY ./digital_asset_types /rust/digital_asset_types
 COPY ./integration_tests /rust/integration_tests
