+++ backfill/Cargo.toml
@@ -1,21 +1,27 @@
 [package]
-name = "program_transformers"
+name = "das-backfill"
 version = { workspace = true }
 edition = { workspace = true }
 repository = { workspace = true }
 publish = { workspace = true }
 
 [dependencies]
+anyhow = { workspace = true }
 blockbuster = { workspace = true }
 bs58 = { workspace = true }
-cadence = { workspace = true }
-cadence-macros = { workspace = true }
 das-core = { workspace = true }
+solana-client = { workspace = true }
+borsh = { workspace = true }
 digital_asset_types = { workspace = true, features = [
   "json_types",
   "sql_types",
 ] }
+anchor-client = { workspace = true }
 futures = { workspace = true }
+clap = { workspace = true }
+log = { workspace = true }
+solana-program = { workspace = true }
+program_transformers = { workspace = true }
 heck = { workspace = true }
 mpl-bubblegum = { workspace = true }
 num-traits = { workspace = true }
