+++ core/src/lib.rs
@@ -1,9 +1,11 @@
 mod db;
+mod metadata_json;
 mod metrics;
 mod plerkle_messenger_queue;
 mod solana_rpc;
 
 pub use db::*;
+pub use metadata_json::*;
 pub use metrics::*;
 pub use plerkle_messenger_queue::*;
 pub use solana_rpc::*;
