+++ nft_ingester/Cargo.toml
@@ -14,6 +14,7 @@ cadence = { workspace = true }
 cadence-macros = { workspace = true }
 chrono = { workspace = true }
 clap = { workspace = true, features = ["derive", "cargo"] }
+das-core = { workspace = true }
 digital_asset_types = { workspace = true, features = [
   "json_types",
   "sql_types",
