+++ nft_ingester/src/tasks/common/mod.rs
@@ -2,10 +2,10 @@ use {
     super::{BgTask, FromTaskData, IngesterError, IntoTaskData, TaskData},
     async_trait::async_trait,
     chrono::{NaiveDateTime, Utc},
+    das_core::{DownloadMetadataInfo, DownloadMetadataNotifier},
     digital_asset_types::dao::asset_data,
     futures::future::BoxFuture,
     log::debug,
-    program_transformers::{DownloadMetadataInfo, DownloadMetadataNotifier},
     reqwest::{Client, ClientBuilder},
     sea_orm::*,
     serde::{Deserialize, Serialize},
