+++ ops/src/bubblegum/mod.rs
@@ -1,6 +1,4 @@
-mod audit;
 mod backfiller;
 mod cmd;
-mod tree;
 
 pub use cmd::*;
